Update patch set 2
Patch Set 2: Code-Review-1

Just re-read all these configs and we need to define a volume size on all of them for the cloud you're in. To use the same volume size as we've got on the current builder you need to add

VOLUME_SIZE=10

That's a 10GB volume. So, if you have an idea of how large a disk you're going to need on these instances, this is how we get to tweak it.

Patch-set: 2
Label: Code-Review=-1