Update patch set 5
Patch Set 5: Code-Review-1

Upgrading a submodule and making large changes is something that we do not allow (or should not be allowing). It makes it hard to pinpoint problems if they come up. The only changes that we should see when a submodule is updated is related to required (or deprecated) code changes to support the new module.

Please rebase off of latest master to pick up change I2f3db21a2bfcd3f895f370bb902067b71b9a00ff

Patch-set: 5
Label: Code-Review=-1