Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Sure no problem, I will move the shell script to the RMR repository.
> 
> Can I pass environment variables to a shell script launched in this way?

Scripts launched this way will have access to all of the ENV vars currently defined at the job step that they are executed in.

If you are wanting to transliterate JJB vars into the scripts this can be done, using !include instead of !include-raw or !include-raw-escape but you're going to need to add in some ShellCheck disable statements to get around problems with that.

Patch-set: 2