Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

Got a spelling mistake in a comment.

Moving the shell script over to the repo itself works, but is really only useful if that makes sense to your code for people outside of a CI as well. My suggestion was to just crack it out to a jjb shell include and hopefully have it generic enough that other projects in our CI could use it.

Patch-set: 2
Label: Code-Review=-1