Update patch set 1
Patch Set 1:

Verification of this change will fail until https://gerrit.o-ran-sc.org/r/c/ci-management/+/342 is merged, then should succeed.

Unfortunately I could not set the DRAFT flag on this change set, that feature is not enabled in this repo and/or this Gerrit. I opened ticket 76874 in RT

Patch-set: 1