Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> David I especially wanted you to see my change in Alpine. It was pinned to python 3 (which could easily jump to 3.8), but now I pinned to python 3.7.  I also changed the container tag you added because it was NOT pinned to version 3.7.3.

agree on removing the '3.7.3' part (i just did that to reflect what the current version in alpine was), but i don't know whether there's any reason to pin to 3.7 as opposed to 3.  are we that worried about breaking changes being introduced in python point releases?  not something i feel particularly strong about, though, so if you prefer then i won't carp about it.

Patch-set: 2