Repin NNG refs to capture proxy reconnect fix
A fix to address NNG issues when reconnecting a session
through a proxy was made on August 9, 2019 and merged into
the NNG master branch with commit ID e618abf8f3db2a9. This
change repins the references to the NNG repo such that the
fix is incorporated in the NNG library.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I84d707b6ac65af8e7d78fa6814208532603ea177