Update patch set 6
Patch Set 6:

I see the build of the ubuntu16 base image fails due to a transient network error.  The minion image is ubuntu also, and it MAY need to be updated so that the base python is 3 (not 2).

For now I'll keep retrying, because some of the builds succeed.

04:59:26     raise ProtocolError('Connection broken: %r' % e, e)
04:59:26 ProtocolError: ("Connection broken: error(104, 'Connection reset by peer')", error(104, 'Connection reset by peer'))

Patch-set: 6