Update patch set 2
Patch Set 2:

A -1 without a constructive suggestion for improvement is not much value to me.  This binding module requires a library.  The library is versioned.  This job needs to get the library and install it.  Right now there is no magic way to get the latest version. Please suggest.

Patch-set: 2
Label: Code-Review=0