Update patch set 2
Patch Set 2:

my proposal is to have a .rmr_version.yml file in the repos that need rmr installed for unit tests to run, and to pull the version from that. This way, when someone makes a commit to their repo, they need not be surprised and/or horrified later when some jjb thing fails, since most people never look into this world.

Patch-set: 2