Revert "Add global read the docs jobs"
This reverts commit 6d1913cfdbc909207af36d0feae26c50ff14b425.
Reason for revert: global RTDv2 jobs fail which blocks work.
The job does not seem to run in the docs/ subdirectory of the repo
that is the target of the change set.  Also the regexps match any
change to any conf.py file, but Python projects have those, not 
necessarily a doc change.

Change-Id: I48a9514515c41287b59e751448037112fd2bfb3c
Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>