Update patch set 1
Patch Set 1:

Thanks, I found the 00-Empty view today, but I've never seen an "All" view in the ONAP or Acumos Jenkins production servers.  Equally confusing to me, I could not find an All view defined for the ONAP or Acumos Jenkins *sandbox* servers, but it is there, staring me in the face.  Just wanna check, are you 100% sure of this?

Patch-set: 1