Update patch set 1
Patch Set 1:

This adds a periodic build to com/log to meet Thoralf's request for regular builds to detect problems in third-party software and dependencies.  It runs every weekend.  Lusheng and I discussed and felt that a daily frequency was a little too much because we cannot establish a reliable notification strategy for failed jobs :/

Patch-set: 1