Update patch set 2
Patch Set 2:

We have some very short lived branches, designed to carry future work while master is treated as frozen, which will go away again in a couple of weeks. And there is no difference in the technology or configuration for verifying the code on those branches. Defining the verify job to trigger on all branches is a convenience, we don't have to add/remove/modify branch lists on a weekly basis.

Patch-set: 2