pti-rtp: change the build timeout to 3000 minutes.
The merge job for pti rtp repo runs out of allocated 9 hours with the
job half done. That implies to double time for building the image,
plus extra time for pushing the image to nexus repo, so increase the
timeout to 3000 to make sure the merge job will succeed. With that
success job execution, we could find out the exact building time and
change this timeout to be more accurate accordingly.

Evidence: https://jenkins.o-ran-sc.org/view/pti-rtp/job/
pti-rtp-shell-merge-master/5/console

16:49:20 NOTE: Running task 3957 of 7857

Issue-ID: INF-6
Signed-off-by: Jackie Huang <jackie.huang@windriver.com>
Change-Id: I74905a69d9b10895cf4af03c69e7633633e63724