Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I need to ask, why increase the timeout by a factor of 5?  Previously 10 hours, now this is 50 hours.  I was told previously that the Yocto jobs run for about 10 hours, about 600 minutes.    Please note that we as non-admin users have no ability to abort a job.

Hi Chris, the job was time out with building progress 50% done. I would expect it could finish in 20 hours. It takes more time than what it did over bare metal server.  the merge job will push the 4G ISO image to nexus repo as well, It's not clear yet how long it will take.  Hence we want to extend the build time long enough to make sure it will not fail due to timeout.  after the first time success we could figure an appropriate build time for this job. 
we are hurry to catch up with the amber release so we submitted this patch for review. 
thanks

Patch-set: 1