Update patch set 2
Patch Set 2:

Well I tried the jjb-deploy trick on this review, then I ran the merge job in the sandbox.  But the merge job could not find my change set, don't know if I put in a bad refspec or if the job is just hardwired to a branch name "master" etc. So the sandbox jobs didn't help.

Patch-set: 2