Skip build of dashboard docker image in Sonar job
No Docker image is needed for Sonar analysis.  The current Sonar job
cannot build a Docker image because it does not login at the registry.
Previously I used goal of "test" to stop the maven build before it
reached the install phase that builds a Docker, but that phase does
not call the jacoco-maven-plugin report goal.

Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>
Change-Id: Iaa924b41eb6d0034bdfacddec490d096e38f0a32