Skip build of dashboard docker image in Sonar job
Drop the override mvn-goal of 'test' and add maven parameter to tell
the Fabric8 maven plugin to skip building the Docker image because
no image is needed for Sonar analysis.  Previously I used goal "test"
to stop the maven build before it reached the install phase that builds
a Docker, but that phase does not call the jacoco-maven-plugin reporter.

Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>
Change-Id: Iaa924b41eb6d0034bdfacddec490d096e38f0a32