Update patch set 1
Patch Set 1:

Oh now I read your comment again -- so the centos node lacks package_cloud but the ubuntu node has it.  That's just weird.  What installs package_cloud to one but not the other?  If possible, if it's not crazy, I'd like to keep them similar.

Patch-set: 1
Label: Code-Review=0