Update patch set 1
Patch Set 1: Code-Review+2

I can't comment on any previous "scoldings," but in the case where the change is directly related to something that global-jjb introduces/fixes, it's ok to include other things along with a global-jjb upgrade. And in this case, it was indeed a mistake (on my part) to use "public" floating IP pool, so that really needs to be fixed.

Patch-set: 1
Reviewer: Gerrit User 1000079 <1000079@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=+2