Update patch set 4
Patch Set 4:

Chris, I had to test a few things to find one that worked. It's not as elegant of a solution as I had imagined, but the way variables/parameters work in JJB is a bit of a mess. But this will enable a single "lf-sonar-prescan" builder that can work for all jobs. If you see any improvements you'd like to make, feel free to comment or submit them.

Patch-set: 4
Reviewer: Gerrit User 1000079 <1000079@84e4e190-e623-4c37-9eda-58b1937c1b53>