Update patch set 1
Patch Set 1:

As submitted here, CI for this new golang project uses these existing job templates:

- Docker verify (compile code, test code, create docker)
- Docker merge  (compile code, test code, create docker, push docker)

I would like to add this template, newly merged to global-jjb and available in ORAN:

- Generic sonar (compile code, test code, analyze code, push report)

This depends on unit tests to exercise the code and measure coverage.  I believe, but am not 100% certain, that the same "go" build and unit-test steps should be done for all three jobs mentioned above.  To avoid duplication/divergence of code, I think the best design for this would be to create a build script for the core go build and test steps, add that script to your git repo, and invoke the script from the Docker *and* from the generic Sonar jobs. That script should call the go-acc tool to run tests AND measure coverage accurately.

I can show you an example of this I worked up for a different repo, if that helps.

Please let me know, thanks.

Patch-set: 1
Reviewer: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>