Update patch set 8
Patch Set 8:

Please do not merge. This change places a LARGE build script in the ci-management repo, which is not desirable. The script that's here should be just a one-line file that calls a main build script, which should be in the e2mgr repository.

Patch-set: 8
Reviewer: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>