Update patch set 3
Patch Set 3:

Well, it turns out that the problem with this solution is that I'm replicating what the Ansible role that installs Go is already doing. This change just creates ANOTHER script in profile.d that isn't being run. So the question is: why does the Jenkins user run the /etc/profile.d scripts on Centos, but not on Ubuntu?

Patch-set: 3
Reviewer: Gerrit User 1000079 <1000079@84e4e190-e623-4c37-9eda-58b1937c1b53>