Update patch set 3
Patch Set 3:

I am against that change because it's not really pythonic. 

If I have to follow it for CICD to work, I'm not willing to die on the hill, but A1 does not follow that (but is not PYPI).

Do I need to follow that for PYPI to work?

Patch-set: 3
Reviewer: Gerrit User 1000074 <1000074@84e4e190-e623-4c37-9eda-58b1937c1b53>