Update patch set 3
Patch Set 3:

Using a subdir is a compromise. It seems to me that the LF gerrit is not pythonic, nor is requiring a docs/ area that pushes to ReadTheDocs.org. By using a subdir you avoid some hassles due to LF choices for CI. But I can only lift a lamp to show you the way, you must choose the fork in the road.

Patch-set: 3