Update patch set 1
Patch Set 1:

with the PR you did into the cicd for a1, there was a cooresponding pr into a1. with the PR you did for the cicd for rmr, there was also a pr into rmr. So in a similar fashion I assumed the py xapp frame was not perfectly set up for this job either.

Patch-set: 1
Reviewer: Gerrit User 1000074 <1000074@84e4e190-e623-4c37-9eda-58b1937c1b53>