Update patch set 2
Patch Set 2:

Definitely a chicken and egg situation.  It's very helpful for the first commits on a project, both the first CI jobs and the first source, to be in flight at the same time because that allows testing verify jobs in the sandbox.  Once all checks out, I try to get the CI merged before the source, so the source gets the real +1 from a verify job.

Patch-set: 2
Reviewer: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>