Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> The image name "bldr-alpine3" is not nearly as descriptive as I would like. It seeks to be all things to all people, bringing gcc, go 1.13, python3, NNG *and* RMR.
> 
> The bldr-alpine3-go image is basically the same but without python3. Is that a key differentiator? I don't think it's any good as a base running (i.e., scratch) image; you don't want gcc and golang in the "from" of a running image.

it's the same after this change you made, but that wasn't true before. 

agree that the original is more of a "builder image" designed for 2 stage, whereas the newer one I created to be used directly as a base, however I don't need it and can use a two stage. 

so this whole thing is a question of whether people use these as builders or bases.

Patch-set: 4
Reviewer: Gerrit User 1000074 <1000074@84e4e190-e623-4c37-9eda-58b1937c1b53>