Update patch set 2
Patch Set 2:

Dave, sorry I bumped the commit message to add the issue ID and lost your vote.

Scott assures me that lib RMR works as before.  And people will be forced to upgrade to a recent RMR soon.  Finally  based on the key error that Tommy identified it's extremely unlikely that anyone uses this base image. So I'd like to move ahead with this version upgrade in the base image.

Patch-set: 2