Update patch set 1
Patch Set 1:

For example A1 uses this in 3 separate docker containers as a base for integration tests. 

The only purpose of this change was to update the version of rmr. 

Alternatively, please go fix all the usages of this container, to use the other image and 2 stage builds, starting with the four Dockerfiles in A1. 

I am blocked By this change since 10am this morning.

Patch-set: 1
Reviewer: Gerrit User 1000074 <1000074@84e4e190-e623-4c37-9eda-58b1937c1b53>