Update patch set 9
Patch Set 9:

(1 comment)

I hope we can test this in the sandbox soon.  It's difficult to get the settings exactly right when the files are in a subdir, the dockerfile and container-tag are in a sub-sub-dir, etc.  Hope this helps.

Patch-set: 9
Reviewer: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>