Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> Please tell me, will this repo have another Docker project like the a1 simulator, or is that simulator the only one that will be maintained here?  If there will be many Docker projects, then I highly recommend adding the gerrit trigger file paths I mentioned in a previous comment. That prevents docker jobs from being triggered inappropriately.  But if there will be only one Docker project, altho not necessary, it's still a good idea because it prevents jobs running when you submit a documentation or other non-code change.

Just to double check, if the docker-root is used, where will the docker build command runs? at near-rt-ric-simulator/ or the repo root? because this will affect the simulator docker file(it needs to be in a line with the docker build command's relative directory to copy some files into container). 

I suggest to merge this commit for now. If there is a need to have multiple docker project in the repo. A update commit will be update in a line with your comments.

Patch-set: 10