Update patch set 10
Patch Set 10: Code-Review+1

If you use docker-root, the build system changes the working directory to that and then invokes Docker.  I see that the paths in the Dockerfile assume that docker is run in the repo root so this is consistent and should work, if we can only get the LF to solve the problem of Docker registry credentials.

Patch-set: 10
Reviewer: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=+1