Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> (1 comment)

Thanks for your feedback. I apprechiate all this. I did't worked on jenkins jobs till now. But regarding the concept and the comment on the release "job". Regarding the point that they do not need "maven settings" I wondered if the approach is woring. The plan is to have "distributions" and "features/bundles" and "parents" in one repo. The "features" parts that is now provided should compile the bundles and uses the sources and pom files in "[oam]/features". So the "-f features" option for maven is also necessary for the release build. Is this a problem?

Patch-set: 2