Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> The maven-settings file has the credentials needed to push docker images to the LF registry.  Don't remove it.
> 
The thing is that when I ran a recheck on the latest commit in the a1-simulator repo, the job failed because it couldn't find the maven-settings file. So I wonder then if something else is wrong and needs fixing?

> Why are you removing the sonar job?  I mean, feel free to remove it, but at some point the integ team (which I'm a member) will be nagging you to put it back.

We had an internal discussion and decided that since this is not production code, we didn't think we needed to run Sonar on it. It is mostly generated code, and very little hand written code at that. But if Sonar is needed anyway, then we will of course keep it and make it work properly.

Patch-set: 1