Update patch set 2
Patch Set 2:

it seems to use this the CICD yamls for projects need to be carefully examined for which build nodes various jobs run on, and pick which job they need based on the build node?

would be cool if there was some kind of "uname" magic that could have been done to detect the OS so that every user doesn't need to do this detailed inspection since people may not even know what particular builder is the default for a stage.

Patch-set: 2