Update patch set 2
Patch Set 2:

(3 comments)

All of my gripes are about copyright and licenses.  Some not updated with 2020, Nokia missing in one, and a yaml file without one.  Noted in comments.    If all are legit, ack and move forward (consider my 0 a +1 and a reason for not -1).  If you fix, assume I'll +1 (I'm going off for dinner and don't want to hold up waiting for non-code concerns) and merge if you wish.

Patch-set: 2
CC: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>