Update patch set 5
Patch Set 5:

> Patch Set 2:
> 
> Based on what I'm seeing others like Ericsson do in open source, when I write a file I only assert AT&T copyright.  
Normally I would agree. However, for reasons not clearly explained, we were instructed to add both companies to the copyright; at least with the initial code pushes in 2019 (all of RMR has dual company notation, yet I am the sole author).  My assumption was that in the negotiated "co-create" agreement everything contributed by one side or the other would carry both.  Then, we were told to only ad AT&T on xAPP code, and the waters became muddy.    

> 
> I have not asserted a license in ANY yaml file within the ci-management repo.
I don't view YAML is something that can carry a copyright, though I can think of an argument for it, I don't see that it applies here.  I was only pointing out that some had one and one did not.

> 
> I could def change 2019 to 2020 in the macro file if you like.
Personally I think leaving it at 2019 is fine; I've been called on the carpet for not adding a range. 

A copyright is good for 75 years, so IMHO none of this will last that long so it matters very little. Further, also IMHO, a range only applies to a top level license as that covers multiple files which are assumed to have copyright years somewhere in that range.  Unlike a patent where adding value to it, and being awarded an extension, is significant because of the relatively short enforcement period, the only thing that is important for a copyright is the initial date.   

I'm good with everything, moving to the current patch set and will +1.

Patch-set: 5