Update patch set 3
Patch Set 3:

The plan is to commit the restructuring commit even though the verify job is failing due to the moved files. Then when that commit is in I will run this job in the sandbox to verify that it is ok.The reason for this plan is that my understanding is that the sandbox job will get the latest from the master branch, it cannot run for a specific change that is only in Gerrit.

Patch-set: 3