Align CI/CD jobs with code refactoring.
Align the CI/CD jobs for the sim/o1-interface project, which has gone through a folder and code refactoring.

The new docker jobs were tested in the Sandbox Jenkins. The verify works fine, the merge jobs are failing because of: unauthorized: access to the requested resource is not authorized

Issue-ID: SIM-19
Change-Id: Ib783ace591499cc1c765a896388aa69cc0357c39
Signed-off-by: Alex Stancu <alexandru.stancu@highstreet-technologies.com>