Add trigger event on package verification and merge jobs for it/dep repo
Currenlty when a commit to it/dep repo is incompatible to the submodules, the packaging verify job will fail.
A trigger event on the package-tag.yaml file is added to make sure that package jobs are only carried out when
package version advances.

This commit also adds packagecloud release jobs for it/dep packages.

Signed-off-by: Zhe <zhehuang@research.att.com>
Change-Id: Ic5f6eedf5cd35b169cac6c8611b66dce5e53961a