Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Felix, it's ok for a job to fail, you can remove a -1 verify failed vote manually and add +1 again manually. I think it's silly to change this config, it just has to be changed right back.

In the case of it/dep, actually it doesn't really make too much sense to do any packaging job (verify or merge) for every commit. This is because it/dep sometimes will have changes that does not work with the ric-plt/ric-dep codes. We need to merge those changes and let ric-plt/ric-dep to catch up.

As a deployment solution, the it/dep package should make sure that the it/dep codes and ric-plt/ric-dep codes are synced up. Therefore, we should really only carry out package verify and merge when we update the ric-plt/ric-dep submodule pointer.

I think this flow is not clear to me yet. We need to brainstorm it together. But for now, at least we should disable the packaging job which is sending out a wrong verify -1 score.

Patch-set: 1