Update patch set 4
Patch Set 4:

You might not have permissions to run jobs in the sandbox - please write a ticket to ask for that.

I launched the job for you, the output is here:

https://jenkins.o-ran-sc.org/sandbox/view/All/job/ric-plt-rtmgr-sonar-prescan-script/1/console

it failed with this error:

06:41:01 + cp -r 'curdir/*' /tmp/go/src/routing-manager/.
06:41:01 cp: cannot stat 'curdir/*': No such file or directory
06:41:01 Build step 'Execute shell' marked build as failure

Please check.

Patch-set: 4
Reviewer: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>