Properly set RMR test args for sonar job
The job to generate sonar data in the RMR repo needs to
only run the unit tests.  This change sets the arguments
passed to make to be just "-R unit" which will execute only
the unit tests to generate the gcov files.  The extra
overhead of building packages and running application tests
is avoided.

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: I9b527d167385c5f389d6793229adc08ef5366424