Update patch set 1
Patch Set 1:

I am not familiar with python specific verify jobs, however I gave it a first shot by using the exact details I noticed in qp-driver. Please see, if this is reasonable attempt or we need to remove any unnecessary stuff. 

Just started with verify job to keep it easy to review. If this is a good start, we can move to adding other jobs as well.

Patch-set: 1