Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> is there a way to do this without embedded spaces? is the -R option required?  The shell script that invokes make is provided by Sonar, so we want to avoid doing anything too weird.

Not that I'm aware of.  The -R option is translated in the makefile to 
   ctest -R unit

which runs only the unit tests.   

From the ctest man page, the only possibilities are:
-R <regex>, --tests-regex <regex>
              Run tests matching regular expression.


It does not provide for --long-arg=value (I tried).

The other option would be to duplicate the make options from the other section in the YAML which builds the package and then runs all of the tests. The advantage of avoiding the application tests is time to generate the packages and time required for the tests them selves (a couple have some delays built in to ensure messages flow).

Patch-set: 1