Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> the make opts seems to accept targets.  Let's try this simple variation:
> 
> -R unit test
> 
> absolutely no quotes anywhere

Don't think that will work as the resulting command will be:
  make -R unit test

and make will attempt to parse/diegest the -R arg and barf.    Make allows variables to be defined on the command line; I assume that CMake sets up ARGS in the generated makefile(s) for consumption as is needed.

Patch-set: 1