Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The simple version gets thru, but the tests fail, please see
> 
> https://jenkins.o-ran-sc.org/sandbox/view/All/job/ric-plt-lib-rmr-c-cmake-sonarqube/1/console

The simple case doesn't present both tokens together.  Make sees "ARGS=-R"  and "unit" and who knows what it's doing with "unit."   Ctest will "match everything" if -R has no trailing value.

What shell is this, or is the quoting being buggered during the yaml parsing?   I've tried various combinations in both Bash and Ksh and I've not been able to make either shell yield the completely buggered quoting.   I'm really guessing that the problem is in the yaml parser which is treating embedded quote as a single quote in the string.   

With that assumption, would this work:

 make-opts: 'test ARGS="-R unit"'

I don't know enough about yaml to know how it treats quotes, but in a shell the single quotes stop the parser from evaluating what is inside.   
The other option, and I don't like it, would be 

 make-opts: test ARGS=\"-R unit\"

Assuming yaml allows escapes.

Patch-set: 1