Update patch set 2
Patch Set 2:

If someone would please find a minute for this, I need a change to the ci-management scripts so I can upgrade A1 to RMR version 3.8.2.  That library is in the release repo, exactly where it belongs, but long ago A1 was wired (incorrectly) to assume that all libraries are in the staging repo.  This change lifts that blocker.  I'd like to produce a new version of A1 to help Ron et al. Thanks.

Patch-set: 2
Reviewer: Gerrit User 1000049 <1000049@84e4e190-e623-4c37-9eda-58b1937c1b53>