Build C-language projects with autotools
First draft of new verify and stage templates with autotools-build.sh
script.  This solution uses GNU autotools to generate a configure file,
generate a Makefile, then build the software.  First trial on the
ric-plt/sdl library.  Uses lf-* file names and lf-* IDs to ease the
task of lifting this to global-jjb.

Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>
Change-Id: I2ccb26b82b64069d6ef95e442c6c76be6a3fd507