New templates for CXX projects using autotools
First draft of verify, stage and sonarcloud job templates with
supporting scripts autotools-build.sh and autotools-sonarqube.sh.
The flow is: run GNU autotools to generate a configure file,
run configure to generate a Makefile, run make to build and test
the code.  For sonar, the project makefile must build test programs,
run the tests, then run gcov to read .gcno/.gcda files and write
.gcov files in the specified report directory.

Uses lf-* file names and lf-* IDs to ease the task of lifting all
to global-jjb.

Test on the ric-plt/sdl library as the first application.

Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>
Change-Id: I2ccb26b82b64069d6ef95e442c6c76be6a3fd507