New templates for CXX projects using autotools
First draft of new verify, stage and sonarcloud job templates with
supporting scripts autotools-build.sh and autotools-sonarqube.sh.
The flow is: run GNU autotools to generate a configure file,
run configure to generate a Makefile, run make to build and test
the code.  Uses lf-* file names and lf-* IDs to ease the task of
lifting everything to global-jjb.

Using the ric-plt/sdl library as a test project.

Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>
Change-Id: I2ccb26b82b64069d6ef95e442c6c76be6a3fd507